requirements.txt
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
defusedxml==0.7.1
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
htmldate==1.9.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
readability-lxml==0.8.4.1
regex==2025.7.34
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
threadpoolctl==3.6.0
tld==0.13.1
tldextract==5.3.0
tqdm==4.67.1
trafilatura==2.0.0
typing_extensions==4.14.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
youtube-transcript-api==1.2.2
